requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.5
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.2
python-dotenv==0.21.1